% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{archprofile_test}
\alias{archprofile_test}
\title{archprofile_test: Fictional fotogrammetrie control points with all direction and view possibilities}
\format{A SpatialDataFrame:
\itemize{
  \item pr: The profile number, respectively the grouping variable
  \item pr1: The direction of view
}}
\description{
A dataset containing four profiles with fotogrammetrie control points
}
\seealso{
Other fotogram: \code{\link{fotogram_pts}}
}
