% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fotogram_pts}
\alias{fotogram_pts}
\title{fotogram_pts: Fictional fotogrammetrie control points}
\format{A SpatialDataFrame:
\itemize{
  \item pr: The profile number, respectively the grouping variable
  \item ansicht: The direction of view
}}
\description{
A dataset containing four profiles with fotogrammetrie control points
}
