% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{KT_spits}
\alias{KT_spits}
\title{KT_data: Niveau measurements from the fictional trench of a excavation KT}
\format{A data frame with 304 rows and 4 variables:
\itemize{
  \item id: IDs of individual measurements with the information about to which level
  they belong
  \item x: x axis coordinates of measurements
  \item y: y axis coordinates of measurements
  \item z: z axis coordinates of measurements
}}
\description{
A dataset containing coordinates of niveau measurements of a fictional excavation KT with
4 spits.
}
\seealso{
Other KT_data: \code{\link{KT_squarecorners}},
  \code{\link{KT_vessel}}
}

