% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{KT_squarecorners}
\alias{KT_squarecorners}
\title{KT_data: Corner points of a 1m*1m raster within the trench of a fictional excavation KT}
\format{A data frame with 63 rows and 2 variables:
\itemize{
  \item x: x axis coordinates of corner points
  \item y: y axis coordinates of corner points
}}
\description{
A dataset containing horizontal coordinates of corner points of a 1m*1m raster within
the rectangular trench (corner points of squares).
}
\seealso{
Other KT_data: \code{\link{KT_spits}},
  \code{\link{KT_vessel}}
}
