% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{KT_vessel}
\alias{KT_vessel}
\title{KT_data: Information about individual sherds of a reconstructed vessel from the trench
of a fictional excavation KT}
\format{A data frame with 7 rows and 7 variables:
\itemize{
  \item inv: Inventory numbers of sherds. KTF means single find with individual measurement,
  KTM means mass find without this precise information.
  \item spit: spits where the sherds were found
  \item square: squares where the sherds were found
  \item feature: features where the sherds were found
  \item x: x axis coordinates of sherds
  \item y: y axis coordinates of sherds
  \item z: z axis coordinates of sherds
}}
\description{
A dataset containing spatial and contextual information for individual sherds of a single
vessel. Some sherds were documented in the field with single find measurements. For the
others only spit and square attribution is possible.
}
\seealso{
Other KT_data: \code{\link{KT_spits}},
  \code{\link{KT_squarecorners}}
}
