% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pnpmulti}
\alias{pnpmulti}
\title{test}
\usage{
pnpmulti(vertx, verty, testx, testy)
}
\arguments{
\item{vertx}{test}

\item{verty}{test}

\item{testx}{test}

\item{testy}{test}
}
\value{
test
}
\description{
test
}
\examples{
polydf <- data.frame(
  x = c(1,1,2,2),
  y = c(1,2,1,2)
)

testdf <- data.frame(
  x = c(1.5, 2.5),
  y = c(1.5, 2.5)
)

pnpmulti(polydf$x, polydf$y, testdf$x, testdf$y)

}

