% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scale}
\alias{scale}
\title{scale}
\usage{
scale(x, y, z, scalex = 1, scaley = 1, scalez = 1)
}
\arguments{
\item{scalez}{}
}
\value{
test
}
\description{
test
}
\examples{
s <- draw_sphere(1,1,1,3)

library(rgl)
plot3d(s)

s2 <- scale(s$x, s$y, s$z, scalex = 4, scalez = 5)

library(rgl)
plot3d(s2)

}

