% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spatialwide2}
\alias{spatialwide2}
\title{spatialwide2}
\usage{
spatialwide2(x, y, z, digits)
}
\arguments{
\item{x}{test}

\item{y}{test}

\item{z}{test}

\item{digits}{test}
}
\value{
test
}
\description{
\code{spatialwide2} test
}

