% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archProfile_function.R
\name{archprofile}
\alias{archprofile}
\title{Tool for rotating profile control points for rectifiying images}
\usage{
archprofile(fotogram_pts, profile_col, view_col, x, y, z, view = "projected",
  direction = "horizontal")
}
\arguments{
\item{fotogram_pts}{SpatialPointsDataFrame or Dataframe. A sp::SpatialPointsDataFrame or dataframe containing the control points (3 Dimensions).}

\item{profile_col}{character. Name of the column containing the profile group variable (profile number).}

\item{view_col}{character. Name of the profile containing the viewing direction ("N","S","W","E").}

\item{x}{character. (optional) If input is a dataframe, column with x coordinates.}

\item{y}{character. (optional) If input is a dataframe, column with y coordinates.}

\item{z}{character. (optional) If input is a dataframe, column with z coordinates.
\itemize{
  \item{"projected"}{: (default) on a vertial intersecting pane like a drawing}
  \item{"surface"}{: orthogonal to the surface of the profile}
}}

\item{view}{character. Direction of view on the Profile.}

\item{direction}{character. Position of the profile control points.
\itemize{
  \item{"horizontal"}{: (default) profile is horizontal, the relative heigth is still correct, good for catalogue export}
  \item{"original"}{:  the points are on the profile section}
}}
}
\value{
sp::SpatialPointsDataFrame with the new coordinates.
}
\description{
Tool for rotating profile control points for rectifiying images
}
\examples{
fotogram_sdf <- sp::SpatialPointsDataFrame(
  coords = fotogram_pts[, c(1,2,3)],
  data = fotogram_pts,
  proj4string = sp::CRS("+proj=utm +zone=32 +ellps=GRS80 +units=m +no_defs")
)

profile <- archprofile(
  fotogram_pts = fotogram_sdf,
  profile_col = "pr",
  view_col = "view"
)

}
