% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archProfile_function.R
\name{archprofile}
\alias{archprofile}
\title{Tool for rotating profile control points form rectifiying images}
\usage{
archprofile(fotogram_pts, profile_col, view_col, view = "projected",
  direction = "horizontal")
}
\arguments{
\item{fotogram_pts}{SpatialDataFrame A SpatialDataFrame containing the control points (3 Dimensions)}

\item{profile_col}{Name of the column containing the profile group variable (Profilenumber)}

\item{view_col}{Name of the profile containing the viewing direction (N,S,W,E)}

\item{view}{Direction of view on the Profile (surface = orthogonal to the surface of the profile, Standard is a view on a vertial intersecting pane (like a drawing)).}

\item{direction}{Position of the profile control points. (horizontal = Profile is horizontal, the relative heigth is still correct, good for catalouge export; original = the points are on the profile section)}
}
\value{
SpatialDataFrame with the new coordinates
}
\description{
Tool for rotating profile control points form rectifiying images
}
\examples{

fotogram_sdf <- sp::SpatialPointsDataFrame(coords = fotogram_pts[ ,c(1,2,3)],
data <- fotogram_pts,
proj4string = sp::CRS('+proj=utm +zone=32 +ellps=GRS80 +units=m +no_defs'))

profile <- archprofile(
  fotogram_pts = fotogram_sdf,
  profile_col = "pr",
  view_col = "view",
)

}
