% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archProfile_function.R
\name{archProfile}
\alias{archProfile}
\title{Tool for rotating profile control points form rectifiying images}
\usage{
archProfile(fotogram_pts, profile_col, view_col)
}
\arguments{
\item{fotogram_pts}{SpatialDataFrame A SpatialDataFrame containing the control points}

\item{profile_col}{Name of the column containing the profile group variable (Profilenumber)}

\item{view_col}{Name of the profile containing the viewing direction}
}
\value{
SpatialDataFrame with the new coordinates
}
\description{
Tool for rotating profile control points form rectifiying images
}
