% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{draw_sphere}
\alias{draw_sphere}
\title{test}
\usage{
draw_sphere(centerx, centery, centerz, r, phires = 10L, thetares = 10L)
}
\arguments{
\item{thetares}{}
}
\value{
test
}
\description{
test
}
\examples{

}

