% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{draw_sphere}
\alias{draw_sphere}
\title{draw_sphere}
\usage{
draw_sphere(centerx, centery, centerz, r, phires = 10L, thetares = 10L)
}
\arguments{
\item{centerx}{test}

\item{centery}{test}

\item{centerz}{test}

\item{r}{test}

\item{phires}{test}

\item{thetares}{test}
}
\value{
test
}
\description{
test
}
\examples{
s <- draw_sphere(1,1,1,3)

#library(rgl)
#plot3d(s)

}

