% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spitcenternat}
\alias{spitcenternat}
\title{spitcenternat}
\usage{
spitcenternat(hex, maplist)
}
\arguments{
\item{hex}{test}

\item{maplist}{test}
}
\value{
test
}
\description{
\code{spitcenternat} test
}

